Consolodated ReportFileNameGenerator and ReportNameGenerator
